Enabled the reverse gear.
The agent can drive backwards, when it is switched on in the code.  To
enable this, comment line 175, "False and" (with a bit of comments behind
that).

The agent will then be able to drive in reverse gear when it has no ammo, and
when self.allow_reverse_gear is True (that is only the case when it is going
for ammo).

This does not work optimal yet!  On average, over 200 games, the agents need
39.1 steps to end the game.  When reverse gear is turned on, it will take
40.82 steps to end the game on average over 200 games. SADFACE